requirements.txt
# Python dependencies for PMP question importer
supabase==2.10.0
python-dotenv==1.0.1
